ai-agents/requirements.txt
# CrewAI Collaborative Agent System Dependencies

# Core CrewAI Framework
crewai>=0.41.0
crewai-tools>=0.4.0

# Language Model Support
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Database and Storage
psycopg2-binary>=2.9.7
redis>=5.0.0
sqlalchemy>=2.0.0
pymongo>=4.6.0

# HTTP and API Integration
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Data Processing and Analysis
pandas>=2.1.0
numpy>=1.25.0
python-dateutil>=2.8.2
pytz>=2023.3

# Task Queue and Background Processing
celery>=5.3.0
redis-py-cluster>=2.1.3

# Configuration and Environment
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Logging and Monitoring
loguru>=0.7.2
prometheus-client>=0.19.0

# File Processing
python-multipart>=0.0.6
pillow>=10.1.0

# Email and Notifications
smtplib-extensions>=0.0.1
twilio>=8.12.0

# Testing and Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
flake8>=6.1.0

# Deployment and Production
gunicorn>=21.2.0
docker>=6.1.0

# Additional ML/AI Libraries
scikit-learn>=1.3.0
transformers>=4.35.0
torch>=2.1.0